+++ pyngrok/exception.py
@@ -1,6 +1,6 @@
 from typing import Any, Optional, List
 
-from pyngrok.ngrok_log import NgrokLog
+from pyngrok.log import NgrokLog
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2023, Alex Laird"